Move Claude secure storage env tests to their own file
